requirements.txt
accelerate==0.31.0
beautifulsoup4==4.12.3
clearml==1.16.2
crowd-kit==1.3.0.post0
datasets==2.20.0
faiss-cpu==1.8.0
huggingface-hub==0.31.2
hydra-core==1.3.2
keybert==0.8.5
matplotlib==3.9.0
networkx==3.3
nltk==3.8.1
numpy==1.26.3
omegaconf==2.3.0
openpyxl==3.1.5
pandas==2.2.2
peft==0.11.1
pymystem3==0.2.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
setfit==1.0.3
tabulate==0.9.0
textfab==1.0.0
torch==2.3.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.39.0
